requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ninja==0.22.2
pip==22.3
pydantic==1.10.12
setuptools==68.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
wheel==0.38.4
